Add files via upload
In this project initially I used the algorithm SVM through which the model attained the accuracy of 96% but then after we improvised the model using the normalization technique by using the "Gamma Prameter Tunning" which uses the complex matrix upgradation to improvise the model and after this we finally got an accuracy of 97% through this model. 
This model is not based on prompts instead this is a complete GUI based website model which can be used by any normal user or patient.